requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
Flask-CLI==0.4.0
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonify==0.5
kagglehub==0.3.12
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyheif==0.8.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
seaborn==0.13.2
setuptools==80.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
thop==0.1.1.post2209072238
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.122
ultralytics-thop==2.0.14
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
